Implementing the NumberMapper that maps a given number to an integer group and back. (#381)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/381

This diff implements the `NumberMapper` class that maps a given `double` number to an integer group.

## Why
The matrix-vector multiplication protocol needs to generate **uniform** random noise to hide the data passed around. Because the input data comes as `double` type, which is essentially unbounded; To generate **uniform** noise we have to first map it to a bounded range.

## How
The arithmetic of `unsigned` type already forms an additive group: all results are automatically modulo `1 << width`. So we will just map the input `double` to an `unsigned` value.
* Template parameters:
  * `FixedPointType`: The target `unsigned` type that will act as the integer group. Can be `uint16_t` to `uint64_t`.
  * `divisor`: Controls how much information are preserved after mapping. An input number will be mapped to the group as **the integral part** of `input * divisor`. One can view `log_10(divisor)` as the number of digits that will be preserved.

## Precision loss
Let `output = mapToUnsignedDouble(mapToFixedPointType(input))`.

Mapping a `double` to the group and then back can cause precision loss:
  * If `input < 0`: Since `unsigned` type only holds non-negative value, a negative `input` will be lost completely. To get correct result, use `mapToSignedDouble` instead.
  * If `input * divisor > (1 << width)`: The group size is too small to hold the desired number of digits. `input` will be lost completely.
  * Otherwise:
     (1). When `input` >= 1.0, the **relative** error will be less than `1 / divisor`.
     (2). When `input` < 1.0, the **absolute** error will be less than `1 / divisor`.
     (3). The error can accumulate when you map many `double` values to `FixedPointType`, sum them up, and then map the sum back to `double`. Users should be aware of this.

Reviewed By: RuiyuZhu

Differential Revision: D39062241

fbshipit-source-id: d6c19175231b84b050c707a38b6b3e7e0179a96f