Change the overflow behavior of `NumberMapper`.
Summary:
## Why
Previously when the `NumberMapper` receives an input that causes overflow (i.e., the input is mapped to a number that exceeds the integer group size), it will log a warning, but proceeds to convert it anyway; It also tries to first convert the `double` input using a widest-possible `intXX_t` to "preserve information".

We note that this is not really useful, since the result will be converted to a `uintXX_t` with possibly smaller width, and the information is lost anyway.
Besides, if overflow happens during the step of converting `double` to `intXX_t`, the behavior is undefined and the result depending on compiler implementation.

Therefore, (in the current implementation) we'd better not specify the overflow behavior of `NumberMapper`, and let the user keep this in mind themselves.

## What

1. When overflow happens, let `NumberMapper` log an ERR rather than WARN, since the conversion result is almost surely meaningless.
2. Remove the overflow unittest, since the behavior is undefined.
3. Define a new estimated value range to detect potential overflow issue. Since we are supporting signed value conversion, the range is now half the group size.
4. Allow the `divisor` to be larger than the group size: previous version will throw an exception for very large `divisor` as an early prevention of overflow, but such `divisor` can be useful for input (abs) value much smaller than 1.

Reviewed By: RuiyuZhu

Differential Revision: D39402511

fbshipit-source-id: ba502a9d2033f2006ca0246a6f033074217d1702