Refactor SocketPartyCommunicationAgentFactory
Summary:
# Why
We have a ton of flaky PCF 2 tests due to networking issues. We need to actually fix them, otherwise they will constantly be in a state of flakiness and won't be able to adequately catch issues.

# What
First, we are refactoring a couple things to make the fix easier. The fix is in the next diff.

Differential Revision: https://internalfb.com/D39493630

fbshipit-source-id: 17a5427aee18894815665d0dc1ee9bd5686922d0