Finally fix flaky tests using an actually guaranteed free port
Summary:
# Why
We have a ton of flaky PCF 2 tests due to networking issues. We need to actually fix them, otherwise they will constantly be in a state of flakiness and won't be able to adequately catch issues.

Right now, we rely on one "guaranteed free port". But this still has a small bug. When running a stress run, these tests run concurrently, which means there is a race condition in our logic when determining a guaranteed free port. The ONLY way to guarantee a free port is to ask the OS to *atomically* bind to a free port and then retrieve it.

So, in this diff, we get a free port from the OS, and then pass it in to the other party to create the connection.

# What
In order to accomplish this, we need to slightly change the way the factory works. Today, the constructor sets up a server port/client port as an "initial connection". We need to tweak this such that the constructor will ONLY bind to a socket but NOT listen to it (i.e. we don't want the constructor to block).

So, the new flow looks as follows
1) Create a socket factory for the publisher
2) Retrieve the port(s) that it bound to
3) Put these ports into a new map of partyInfos
4) Create a socket factory for the partner with this new information

To accomplish this, we create a duplicate version of the socket factory that exposes the necessary API instead of changing the production API. This new version will only be needed for tests.

# Next
We will put up another diff to make this change for the PCF2 Calculator, Shard Combiner, and Attribution app tests.

Differential Revision: D39493637

fbshipit-source-id: aa19bc9b87cbb29bd89efd068722c31cd5fd10c2