Refactor SocketPartyCommunicationAgentFactory
Summary:
# Why
We have a ton of flaky PCF 2 tests due to networking issues. We need to actually fix them, otherwise they will constantly be in a state of flakiness and won't be able to adequately catch issues.

# What
First, we are refactoring a couple things to make the fix easier. The fix is in the next diff.

Differential Revision: D39493630

fbshipit-source-id: a51d5a0f14ffc2b23329fc6723287b17887eb1e2