Refactor SocketPartyCommunicationAgentFactory (#394)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/394

# Why
We have a ton of flaky PCF 2 tests due to networking issues. We need to actually fix them, otherwise they will constantly be in a state of flakiness and won't be able to adequately catch issues.

# What
First, we are refactoring a couple things to make the fix easier. The fix is in the next diff.

Reviewed By: RuiyuZhu

Differential Revision: D39493630

fbshipit-source-id: 3cb085c0def5fd73deef19a0a9c8cda27011ff3f