Record metrics for composite tuples (#401)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/401

We add metrics in the TuplesMetricRecorder for composite tuples.

Add the following metrics to the TuplesMetricRecorder:
number of RCOTs used (1-1 mapping to number of composite tuples requested)
total of mini tuples requested with size <= 128
total of mini tuples requested with size > 128
number of composite tuples requested with size <= 128 (no need for AES expansion)
number of composite tuples requested with size > 128
```
Example: if 3 request comes with sizes = 20, 40, 150. Then the metrics are
(3, 60, 150, 2, 1)
```
Verified the collector works correctly in TupleGeneratorTest.cpp

Reviewed By: RuiyuZhu

Differential Revision: D39452273

fbshipit-source-id: 21f0cfa33f6451622723acda52e81498e208570d