Re-write testWithJammedPort to be robust (#402)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/402

This is another flaky test. It is also relying on a non-atomic version of determining a free port, which will fail stress tests.

Instead, I have rewritten the test to create a guaranteed set of agent factories (set up in a previous stack of diffs). When we create agents, it will automatically try to connect to port 1. This isn't allowed, so it should test our functionality. Then, I create agents and ensure it doesn't time out.

Reviewed By: nguytc

Differential Revision: D39682065

fbshipit-source-id: 55ac66fcb0adbc810c3757a4b7d2d82163ed9753