Add methods to record traffic statistics
Summary:
## Why
The goal is to support micro-benchmark for the matrix-multiplication method in subsequent diffs. Therefore we need a way to report the network traffic.

## What
* Add `getTrafficStatistics` method for the `IWalrMatrixMultiplication` class and all its descendants.
  * The traffic has two possible sources, one is the MPC engine (since we accept an `frontend::Bit`-type label vector), the other is the traffic from the matrix multiplication protocol itself.
  * We refactored the `IWalrMatrixMultiplication::matrixVectorMultiplication` method to let it measure the engine traffic difference before and after running the matrix multiplication.

* Correct the `getTrafficStatistics` method of `COTwithRandomMessage` to include the traffic from its underlying RCOT.

## Note
Some offline tests show that, the current two implemented matrix multiplication classes (i.e. `DummyMM` and `OTBasedMM`) have minimal to zero engine traffic. This is expected, since the two do not conduct any arithmetics on the secret shares.

Reviewed By: chualynn

Differential Revision: D39527971

fbshipit-source-id: 4b944cfb7891b3b24476ed1521f54693805bc054