Adding timeout on client side (#413)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/413

# Why
See last diff, this is a follow-up on S291746

# This diff

This diff add a timeout mechanism when creating a connection as the client.
The client will try to connect to the given server repeatedly. it will retry after a wait if an attempt failed. This diff specifies that the wait time is 100ms and number of retry is limited to 10 * timeout_time_in_secs.

Reviewed By: adshastri

Differential Revision: D39562698

fbshipit-source-id: 8c24b560e97f96e8e2a99cf4557aa3c99860cae2