add a secure seed sampling algorithm (#424)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/424

As title. This function will be used to sharding secret-shares randomly.
The algorithm is straightforward.
The public seed will be the XOR of two random strings from both parties. The two parties exchange the random strings after one of them commit to their own random string first. The commitment is done via a salted hash.
The protocol is summarized below.
1. party 1 and party 2 sample a random string s1 and s2 respectively.
2. party 1 computes hash(s1||r) and share the hash with party 2. r is a random string as well.
3. party 2 shares s2 with party 1.
4. party 1 shares s1 and r with party 2.
5. party 2 verifies s1 and r against the hash received before.
6. the two parties outputs s1 xor s2.

Reviewed By: robotal

Differential Revision: D40326612

fbshipit-source-id: 97b59f2a80e688e9fccfc4fd59e477b338595d6f