Add a Hotfix workflow for FBPCF
Summary:
## Background
As of right now, there is not a good way for us to Hotfix any changes to FBPCF. It is a continuous release process and if there is a bug, the standard process right now is to fix the bug, land the change, and then bump the version. This presents an issue in the case where a given release (say 2.2.0 for example) includes a bug which has been deployed to Canary or Production. This could cause a SEV and lead to us needing to release a new hotfixed release. If 2.2.1 has already been released by the time that we figure out 2.2.0 has a bug and 2.2.1 also introduces a different unrelated bug, this means that we need to wait for resolutions* to the 2.2.0 bug and the 2.2.1 bug in the current flow before we can release 2.2.2 and update FBPCS. We also may be under a code freeze and want to release as small of a fix as possible to production, which means we either can’t land any new code to FBPCF after the last release of the year, or we need another solution.

## This diff
This diff adds a new workflow that can be used to create a hotfix release. It is intended to be used to apply fix commits directly on a release version that is in prod. It does this by creating a new branch off the base tag (v2.2.0) and cherry-picking a comma separated list of commit hashes onto the hotfix branch. It then builds and tests the resulting docker image and creates a release in the format of 'v{base_tag}.{commits_since_the_base_tag}'. From the example above, if there were one commit to cherry pick, the resulting version would be 2.2.0.1.

This workflow would also support a second hotfix to the same tag. In the event of a long lived release, we may find more than one issue that needs hotfixed. By inputing the same base tag, you are able to check out the existing hotfix branch and add the fixes to the branch.

Differential Revision: D40858259

fbshipit-source-id: 3f040a3b87b4c4c9983456d069b6043f856fe228