Create new target for `io` apis and its test and delete monolith target (#437)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/437

# Context
We want to refactor the targets in fbpcf to follow the convention of one folder, one TARGETS file.

# This Diff
Create new target for the `io` folder and its children. It also deletes the monolith target and removes all of its usages.

# This Stack
This stack attacks the problem folder by folder, and removes pieces from the monolith target `lib_fbpcf`

Differential Revision: D40728409

fbshipit-source-id: a62966ef63524cb078cbc84a1f218af49bbb8568