Create new target for `io` apis and its test and delete monolith target (#437)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/437

# Context
We want to refactor the targets in fbpcf to follow the convention of one folder, one TARGETS file.

# This Diff
Create new target for the `io` folder and its children. It also deletes the monolith target and removes all of its usages.

# This Stack
This stack attacks the problem folder by folder, and removes pieces from the monolith target `lib_fbpcf`

Reviewed By: tgamal

Differential Revision: D40728409

fbshipit-source-id: fbfb1f6baab028064c95a8439a9578080e27ffb4