Fix FBPCF Hotfixing
Summary:
## Context
After a long search I finally found https://github.com/orgs/community/discussions/25702 which explains why when we run the "Create Hotfix" action, the subsequent push-based action is not triggered. It recommends using a personal access token, but I don't think that is a good idea for our use case as anyone whose token we might use could eventually be removed as a contributor and then our pipeline would stop working. The next best option is adding a deploy key. This is essentially an SSH key that can be used by anyone to deploy the repo. After it is added, I can update the checkout action to use that deploy key to push the hotfix and trigger the hotfix deployment.

## This diff
This diff updates the create hotfix workflow to use the new deploy key that will be created as part of T136993451. After that, the hotfix workflow should work as intended.

Differential Revision: D40987028

fbshipit-source-id: 6e70839d5f3cbf2fc70ea2c5e79b6bf7bb737495