attempt to improve permuter efficiency (#445)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/445

Try to rewrite the permuter for better performance by adding a specialization for string type.

Before:
{F788674913}

After:
{F788674914}

The idea is to manually "schedule" the gates such that there is only 1 gate per layer and the burden of scheduling, "batching/rebatching" are omitted.

Based on the numbers it looks like we have saved roughly 76% percentage of the latency.

The algorithm could apply to other data types but would require customized serialization/deserialization. This could be a BE project.

Reviewed By: robotal, adshastri

Differential Revision: D40924928

fbshipit-source-id: 9e6361aaab02f13a01b364e7c1bafc0190d5ea79