Add argument validation for unionMaps (#467)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/467

In T139445158 the NCC brought to our attention that the permuter library does not actually verify that the user has provided a valid permutation (i.e. same size as dataset, each index from [0 to n-1] appears exactly once. This diff adds that input validation.

Note: The other party that calls permute or adapt will just hang and timeout when this happens as it has no way to know the other party aborted.

Reviewed By: RuiyuZhu

Differential Revision: D42462081

fbshipit-source-id: 4f023262a10af012e4105fdbde1f4c1e74e7894b