Update the FBPCF Dependency Docker Images to be built with the Haswell architecture (#475)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/475

## Context
After attempting to build the images with the "x86-64" architecture for a while, I determined that we wouldn't be able to make it work as the EMP library used an instruction that wasing included in the "x86-64" architecture (specifically ). This meant that I needed to explore other options. My next choice was to find the architecture of the current build machines that we use (c4.4xlarge) and go with that since we know that we were building it previously natively for those machines. The c4.4xlarge uses an Intel Xeon E5-2666 based on the cpu info from the machine:
```
ubuntu@ip-172-31-15-42:~$ lscpu
Architecture:                    x86_64
CPU op-mode(s):                  32-bit, 64-bit
...
Model name:                      Intel(R) Xeon(R) CPU E5-2666 v3 @ 2.90GHz
...
```

According to the Intel website, this processor uses the Haswell architecture: https://www.intel.com/content/www/us/en/products/sku/81706/intel-xeon-processor-e52660-v3-25m-cache-2-60-ghz/specifications.html

With that information, I decided to rebuild everything for the haswell architecture since it should match our current build system and allow us all the symbols necessary to run the computation.

## This Diff
This diff updates all of the dependent FBPCF images (AWS, EMP, and Folly) to be built with the haswell architechture specified. After this has been done and confirmed to work. I will produce new haswell images and update the main FBPCF build in the next diff.

Reviewed By: robotal, ramesc

Differential Revision: D42588188

fbshipit-source-id: 18767d04911027139c40125c1833228616ce4939