Update PCF binaries to be built using the haswell archetecture (#476)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/476

## Context
After attempting to build the images with the "x86-64" architecture for a while, I determined that we wouldn't be able to make it work as the EMP library used an instruction that wasing included in the "x86-64" architecture (specifically ). This meant that I needed to explore other options. My next choice was to find the architecture of the current build machines that we use (c4.4xlarge) and go with that since we know that we were building it previously natively for those machines. The c4.4xlarge uses an Intel Xeon E5-2666 based on the cpu info from the machine:
```
ubuntu@ip-172-31-15-42:~$ lscpu
Architecture:                    x86_64
CPU op-mode(s):                  32-bit, 64-bit
...
Model name:                      Intel(R) Xeon(R) CPU E5-2666 v3 @ 2.90GHz
...
```

According to the Intel website, this processor uses the Haswell architecture: https://www.intel.com/content/www/us/en/products/sku/81706/intel-xeon-processor-e52660-v3-25m-cache-2-60-ghz/specifications.html

With that information, I decided to rebuild everything for the haswell architecture since it should match our current build system and allow us all the symbols necessary to run the computation.

## This Diff
This diff updates the FBPCF binaries to use the new "haswell" image versions and specifies the haswell architecture for the FBPCF binaries also.

Reviewed By: ajitfb

Differential Revision: D42627125

fbshipit-source-id: 6954c6f00b767ad789e97f9e2b4273f4a78cbb0b