Split the test cases to avoid timeout errors
Summary:
We have seen a task T145020747 that some tests are flaky because they exceed the 600s limit. This diff splits the test cases to reduce the time needed to execute each test case.

Thanks adshastri for suggesting this fix. We will keep monitoring if this helps reduce flakiness and we will gradually apply it to other test cases

Reviewed By: adshastri

Differential Revision: D43200388

fbshipit-source-id: e41e97a24d9415c7431f1f2864124682cce81dc1