Add column for serializing a packed bit field. (#498)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/498

# Background:

Currently in order to successfully use UDP, you must write some carefully crafted code that will take all the rows of metadata for one side and package it into a collection of bytes. Afterwards the caller will get a `SecString` object back which is a bit representation of all the bytes they passed in, minus the filtered out rows. The user must then extract the corresponding bits for each column into separate MPC Types.  This is a cumbersome process which is error prone, as you must make sure to carefully match up the two steps and any changes can cause a bug.

# This Diff

Adds support for a packed bit field column. The current plan is that the user will not directly interface with this column type, but it is internal to the row structure. It will support up to 8 different boolean values to be packed inside of it. The reason for this is that a bool value takes up a whole byte in memory. This means if we naively store these, each bool will actually take 8 bits in the final row structure. This allows us to fit 8 bool columns into one byte!

Reviewed By: haochenuw

Differential Revision: D43289316

fbshipit-source-id: 3d7fa689a4f672168d302d4d0481278f7a69a81e