Add ColumnDefinition interface (#489)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/489

# Background:

Currently in order to successfully use UDP, you must write some carefully crafted code that will take all the rows of metadata for one side and package it into a collection of bytes. Afterwards the caller will get a `SecString` object back which is a bit representation of all the bytes they passed in, minus the filtered out rows. The user must then extract the corresponding bits for each column into separate MPC Types.  This is a cumbersome process which is error prone, as you must make sure to carefully match up the two steps and any changes can cause a bug.

# This Diff

Each "variable" in the MPC game will be represented by a single `IColumnDefinition` object. It will know how to serialize the data that is passed into it by the caller, as well as how to load it back into the correct MPC Type. We are using `std::variant` to support all the different input and output types.

Reviewed By: RuiyuZhu

Differential Revision: D43366176

fbshipit-source-id: 6a0a7e499834ad499792dbca8d4a7aeaebd1e5e1