Log partner ip address from publisher container (#499)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/499

## what
This diff finds and logs the client's ip address after the socket agent server establishes a connection with client.

## why
To investigate runs, we often have information about the ip of failed publisher container, and this log provides an easy way to find the ip of the corresponding partner container.

Reviewed By: adshastri

Differential Revision:
D43484311

Privacy Context Container: L323792

fbshipit-source-id: fe78f9733d0f7aa065095046b1981f8d1a1956f7