Add padded vector column type (#491)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/491

# Background:

Currently in order to successfully use UDP, you must write some carefully crafted code that will take all the rows of metadata for one side and package it into a collection of bytes. Afterwards the caller will get a `SecString` object back which is a bit representation of all the bytes they passed in, minus the filtered out rows. The user must then extract the corresponding bits for each column into separate MPC Types.  This is a cumbersome process which is error prone, as you must make sure to carefully match up the two steps and any changes can cause a bug.

# This Diff

Adds support for padded array column types. The padding size is passed in at run-time and will determine the size of the column.

Reviewed By: haochenuw

Differential Revision: D43366171

fbshipit-source-id: fb3e90654c4a0b50e45eb66732c3c6c16513ab45