Add column type enum for dynamic mapping (#497)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/497

# Background:

Currently in order to successfully use UDP, you must write some carefully crafted code that will take all the rows of metadata for one side and package it into a collection of bytes. Afterwards the caller will get a `SecString` object back which is a bit representation of all the bytes they passed in, minus the filtered out rows. The user must then extract the corresponding bits for each column into separate MPC Types.  This is a cumbersome process which is error prone, as you must make sure to carefully match up the two steps and any changes can cause a bug.

# This Diff

In order to be able to dynamically create column types at runtime, we need some simple global mapping to all the valid template types we want to support. The idea is that each class will have a unique enum value that we can then use to instantiate it at runtime through a command line structure mapping column names to their types.

Reviewed By: haochenuw

Differential Revision: D43341286

fbshipit-source-id: b785b0530750c4c4556d5bd2dfd9893698c0de24