Add packed bit field support to Row Structure (#493)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/493

# Background:

Currently in order to successfully use UDP, you must write some carefully crafted code that will take all the rows of metadata for one side and package it into a collection of bytes. Afterwards the caller will get a `SecString` object back which is a bit representation of all the bytes they passed in, minus the filtered out rows. The user must then extract the corresponding bits for each column into separate MPC Types.  This is a cumbersome process which is error prone, as you must make sure to carefully match up the two steps and any changes can cause a bug.

# This Diff

Adds support to RowDefinition for packed bits. Right now you directly pass in the column names that are to be packed, and the interface will combine each column. The output will contain columns in a vector form. I will refactor that in a future diff to make a bit easier to use.

Reviewed By: haochenuw

Differential Revision: D43366173

fbshipit-source-id: 6604bad668dbd4acaf7e59d3b60dfc0be217a091