Add RowDefinition implementation with integer column support (#490)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/490

# Background:

Currently in order to successfully use UDP, you must write some carefully crafted code that will take all the rows of metadata for one side and package it into a collection of bytes. Afterwards the caller will get a `SecString` object back which is a bit representation of all the bytes they passed in, minus the filtered out rows. The user must then extract the corresponding bits for each column into separate MPC Types.  This is a cumbersome process which is error prone, as you must make sure to carefully match up the two steps and any changes can cause a bug.

# This Diff

Creates the implementation for the RowDefinition. As a start we only support integer columns, more types to come in later diff.

Reviewed By: RuiyuZhu

Differential Revision: D43208067

fbshipit-source-id: 94a1141233a0a613d3ded6a8975f0ddec0d6da17