Change RowStructureDefinition ctor to be easier to use (#496)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/496

# Background:

Currently in order to successfully use UDP, you must write some carefully crafted code that will take all the rows of metadata for one side and package it into a collection of bytes. Afterwards the caller will get a `SecString` object back which is a bit representation of all the bytes they passed in, minus the filtered out rows. The user must then extract the corresponding bits for each column into separate MPC Types.  This is a cumbersome process which is error prone, as you must make sure to carefully match up the two steps and any changes can cause a bug.

# This Diff

We already have a convenient enum mapping of names to ColumnTypes. Rather than having the user manually create these each time, it is easier to just pass it in as a map from column name to the enum type for that column. I rewrote the test to use the new ctor.

Reviewed By: RuiyuZhu

Differential Revision: D43366174

fbshipit-source-id: d1b5681884f79d3910a3f32f7342eb07bd8fb7bb