Add Integer Vector Support to Row structure (#495)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/495

# Background:

Currently in order to successfully use UDP, you must write some carefully crafted code that will take all the rows of metadata for one side and package it into a collection of bytes. Afterwards the caller will get a `SecString` object back which is a bit representation of all the bytes they passed in, minus the filtered out rows. The user must then extract the corresponding bits for each column into separate MPC Types.  This is a cumbersome process which is error prone, as you must make sure to carefully match up the two steps and any changes can cause a bug.

# This Diff

Adds support for Vector columns to the RowDefinition.

Reviewed By: haochenuw

Differential Revision: D43366175

fbshipit-source-id: 8554b79125205cab35096d33da9423f38163e135