Switch Column serialization functions to pass data by ref (#500)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/500

Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/496

# Background:

Currently in order to successfully use UDP, you must write some carefully crafted code that will take all the rows of metadata for one side and package it into a collection of bytes. Afterwards the caller will get a `SecString` object back which is a bit representation of all the bytes they passed in, minus the filtered out rows. The user must then extract the corresponding bits for each column into separate MPC Types.  This is a cumbersome process which is error prone, as you must make sure to carefully match up the two steps and any changes can cause a bug.

# This Diff

The current path for passing plaintext data to be serialized is slightly error prone. Each column type expects a certain vector of values (i.e. Int32 column would expect a `std::vector<int32_t>` types). Internally the columns definition will take this data as a void* and the implementation will serialize the value after casting it to the value type it expects.

This tightly couples the RowStructure implementation to the data type that is expected as each column will have a different width that the void* should be incremented.

The major change here is to `IColumnDefinition::serializeColumnAsPlaintextBytes`. It now requires a reference to the whole column to be serialized, as well as the vector of writeBuffers for the whole storage operation. We have changed the input type to be a variant of the reference type that should be serialized. This means we have better run-time errors if the wrong variant is passed in by the user (rather than a silent failure / data corruption due to using a void*). Furthermore the data to be passed in no longer needs to be passed by value / copied into the input data map. It is enough to create references to the data for each column, without having them in a single map data structure.

Reviewed By: RuiyuZhu

Differential Revision:
D43410376

Privacy Context Container: L1121121

fbshipit-source-id: 4ac7daba9a4e0fd4f352a7f4e8be91d16ef95452