Fix major serialization bug (#503)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/503

Found this bug while testing PL on Explat. There is an n^2 loop. The column implementation is already writing the values for each row to the byte buffers (https://fburl.com/code/a48hfogh). This effectively means we are processing this column n^2 times. For 1 million rows that is 1 trillion times we are calling this method 🤦‍♀️

Differential Revision: D43954084

fbshipit-source-id: 0ad2e372eacfa61d4331350814cd0629bd126abe