Limit PC run selection to only builds on the main branch
Summary:
## Context
There are times that we run the publish OneDocker job on feature branches in the FBPCS repo to test changes that we are making. This is problematic for the FBPCF repo since they require the last commit from this API to be available on the main branch.

## This Diff
This diff adds the filter that requires the FBPCS run to be on the main branch.

Reviewed By: RuiyuZhu

Differential Revision: D44140323

fbshipit-source-id: 3dd595fcbc256286286344f4ba1d7c997b05a31a