ORAM encoder API definition (#505)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/505

# Context
The ORAM encoder / decoder library is a simple component that will be utilized to generate unique group by keys for ORAM aggregation logic. It supports the following flow
1. Collect all the rows of data being grouped (assuming the plaintext values are known to a single party).
2. Collect Filters to apply to this data.
3. Map each row to a single index in the ORAM step
4. Perform ORAM aggregation.
5. Map the buckets back to their original values. Any values that have been filtered out will decode to an empty vector.

# This diff

Adds the base API for an ORAM encoder. Some of the specific signatures will change a little in future diffs but this is the general idea.

Reviewed By: haochenuw

Differential Revision: D44040849

fbshipit-source-id: b23116f74802a0f929e3a083ccf60d94244e3fd1