Implement ORAM encoding without filters (#506)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/506

# Context
The ORAM encoder / decoder library is a simple component that will be utilized to generate unique group by keys for ORAM aggregation logic. It supports the following flow
1. Collect all the rows of data being grouped (assuming the plaintext values are known to a single party).
2. Collect Filters to apply to this data.
3. Map each row to a single index in the ORAM step
4. Perform ORAM aggregation.
5. Map the buckets back to their original values. Any values that have been filtered out will decode to an empty vector.

# This diff

Implementation of the ORAM encoding functionality. We serialize the inputs into a comma separated string and perform a lookup in the existing map. If there is no match then increment the current index and store the value before returning it.

Reviewed By: haochenuw

Differential Revision: D44040848

fbshipit-source-id: 0c15574dbe9a0d33f719722be14a221aad6278e4