Serialization / Deserialization Support (#509)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/509

Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/508

# Context
The ORAM encoder / decoder library is a simple component that will be utilized to generate unique group by keys for ORAM aggregation logic. It supports the following flow
1. Collect all the rows of data being grouped (assuming the plaintext values are known to a single party).
2. Collect Filters to apply to this data.
3. Map each row to a single index in the ORAM step
4. Perform ORAM aggregation.
5. Map the buckets back to their original values. Any values that have been filtered out will decode to an empty vector.

# This diff

Building upon the previous "export" functionality. We add the ability to serialize the state as a JSON string and to parse it back. This should allow for storage to file system to re-use in a different binary.

Differential Revision: D44041704

fbshipit-source-id: 206155792783dc740b27323468b31c8aa14f3d61