Decoding Support (#508)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/508

# Context
The ORAM encoder / decoder library is a simple component that will be utilized to generate unique group by keys for ORAM aggregation logic. It supports the following flow
1. Collect all the rows of data being grouped (assuming the plaintext values are known to a single party).
2. Collect Filters to apply to this data.
3. Map each row to a single index in the ORAM step
4. Perform ORAM aggregation.
5. Map the buckets back to their original values. Any values that have been filtered out will decode to an empty vector.

# This diff

Adds the ability to "export" an ORAM mapping from an Encoder and "import" it into a decoder. The decoder will perform the reverse operation depending on the input values. It will return an empty vector for the filtered bucket.

Reviewed By: haochenuw

Differential Revision: D44040847

fbshipit-source-id: c31ed410223b845beb8d93c2001b48a099a6af85