change int32_t to uint64_t for ids in UDP in fbpcf
Summary:
As title.

We used to int32_t for index in UDP to save cost (b/c it is used in adapter).
Now as we integrating with i-PID, adapter is no longer need but the index could be larger, we change it to use uint64_t.
In this diff we make the change in fbpcf repo. In the next few diffs we will make the change in fbpcs and remove the temporary APIs in this diff.

Differential Revision:
D44242198

Privacy Context Container: L416713

fbshipit-source-id: bbbef59fe0579ee662b4e04a0802666c37b2dff6