change int32_t to uint64_t for ids in UDP in fbpcf (#516)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/516

As title.

We used to int32_t for index in UDP to save cost (b/c it is used in adapter).
Now as we integrating with i-PID, adapter is no longer need but the index could be larger, we change it to use uint64_t.
In this diff we make the change in fbpcf repo. In the next few diffs we will make the change in fbpcs and remove the temporary APIs in this diff.

Reviewed By: adshastri

Differential Revision:
D44242198

Privacy Context Container: L416713

fbshipit-source-id: c10c4b3d44f977046a364ddff2a38cd2bf652c14