Remove temporary APIs in fbpcf repo
Summary:
As title.

We used to int32_t for index in UDP to save cost (b/c it is used in adapter).
Now as we integrating with i-PID, adapter is no longer need but the index could be larger, we change it to use uint64_t.
In this diff remove the temporary APIs in fbpcf.

Reviewed By: robotal

Differential Revision: D44247962

fbshipit-source-id: 8de77a110065c96ad92c42f280280c867707888f