Remove temporary APIs in fbpcf repo
Summary:
As title.

We used to int32_t for index in UDP to save cost (b/c it is used in adapter).
Now as we integrating with i-PID, adapter is no longer need but the index could be larger, we change it to use uint64_t.
In this diff remove the temporary APIs in fbpcf.

Differential Revision: https://internalfb.com/D44247962

fbshipit-source-id: 107beaed4e258e062773a17226af828abec16dda