update udp encryption API to take in customized indexes in fbpcf
Summary:
We will use 3 diffs to change the udp encryption API to support customized indexes.
This diff makes the change in fbpcf, while maintaining a temporary API for backward compatibility.
In the next diff, we will make the change in fbpcs, which complete the change.
In the last diff, we will clean up the temporary APIs in fbpcf.

Reviewed By: robotal

Differential Revision: D44242200

fbshipit-source-id: 48a36929f27aa84cb4dfe4396f00562764db190e