Remove temporary APIs in fbpcf repo (#517)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/517

As title.

We used to int32_t for index in UDP to save cost (b/c it is used in adapter).
Now as we integrating with i-PID, adapter is no longer need but the index could be larger, we change it to use uint64_t.
In this diff remove the temporary APIs in fbpcf.

Reviewed By: robotal

Differential Revision: D44247962

fbshipit-source-id: 65456ceb46981ca98d45fc9933714d3a23f65a3e