Split the test cases in ProductShareGenerator to avoid timeout errors (#521)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/521

We have previously seen an oncall task T145020747 that some tests are flaky because they exceed the 600s limit.  We have released a diff D43200388  to reduce the likelihood of this timeout issue and we recommend to extend this solution to other tests that call multiple functions (T145644394).

This diff splits the test cases to reduce the time needed to execute each test case in the ProductShareGenerator.cpp file

Reviewed By: adshastri

Differential Revision: D44389601

fbshipit-source-id: c03a05de3fa304bfc7c196d4d6f4bc64cb4e6e1c