Split the test cases for oblivioustransfer to avoid timeout errors (#522)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/522

We have previously seen an oncall task T145020747 that some tests are flaky because they exceed the 600s limit.  We have released a diff D43200388  to reduce the likelihood of this timeout issue and we recommend to extend this solution to other tests that call multiple functions (T145644394).

This diff splits the test cases to reduce the time needed to execute each test case in the BidirectionObliviousTransferTest.cpp file

Reviewed By: adshastri

Differential Revision: D44388804

fbshipit-source-id: 916b072c8f268cc4c6097e38ff94de56b8ca59b0