upgrade AWS CPP SDK from 1.8.177 to a newer version
Summary:
We got a SEV S322350 where the s3 upload showed successful while the file was not uploaded at all. It was due to a bug in 1.8.177 and it was fixed in 1.9.333.

Chose 1.9.379 because it's the last version before 1.10.0 and 1.10.0 has major changes on the s3client.h which isn't compatible with our code anymore.

Differential Revision: D47244697

fbshipit-source-id: 7f3bf8c1a51902e6c09566f28c8ba77593edb543