Remove unused variables in fbpcf/mpc_std_lib/unified_data_process/serialization/RowStructureDefinition.h (#531)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/531

LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D53779568

fbshipit-source-id: 5e551ab8b86f1b1fd3027f1d3b123b11a6a1a46b