Remove unused variables in fbit/itstorage/data_crawler/thread/ThreadedCrawler.cpp (#538)
Summary:
X-link: https://github.com/facebookresearch/fbpcs/pull/2397

Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/538

LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D56022415

fbshipit-source-id: 65aa2b310e39ff134bf6dd8a82e3a539c3033660